add source locale file and drop generated one
